Move tests that are more aligned with fileconnector than the state db
