+++ lucille-core/src/test/java/com/kmwllc/lucille/connector/FileConnectorTest.java
@@ -15,13 +15,17 @@ import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
+import com.kmwllc.lucille.connector.jdbc.DBTestHelper;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.time.Instant;
 import java.util.List;
 import java.util.Map;
 
+import org.apache.commons.io.FileUtils;
 import org.junit.Assert;
+import org.junit.Rule;
 import org.junit.Test;
 import org.junit.jupiter.api.parallel.Execution;
 import org.junit.jupiter.api.parallel.ExecutionMode;
@@ -45,6 +49,9 @@ import com.typesafe.config.ConfigValueFactory;
 public class FileConnectorTest {
 
   private static final Logger log = LoggerFactory.getLogger(FileConnectorTest.class);
+
+  @Rule
+  public final DBTestHelper dbHelper = new DBTestHelper("sm-db-test-start.sql");
 
   @Test
   public void testExecuteSuccessful() throws Exception {
@@ -432,4 +439,237 @@ public class FileConnectorTest {
     assertTrue(sawWeird);
     assertTrue(sawNormal);
   }
+
+  @Test(expected = IllegalArgumentException.class)
+  // Constructing a FileConnector with sendTombstones=true and publishMode=full must throw.
+  public void testSendTombstonesRequiresIncrementalMode() throws Exception {
+    Config config = ConfigFactory.parseResourcesAnySyntax("FileConnectorTest/state.conf")
+        .withValue("filterOptions.publishMode", ConfigValueFactory.fromAnyRef("full"))
+        .withValue("filterOptions.sendTombstones", ConfigValueFactory.fromAnyRef("true"));
+    new FileConnector(config);
+  }
+
+  @Test(expected = IllegalArgumentException.class)
+  public void testSendTombstonesRequiresPublishMode() throws Exception {
+    Config config = ConfigFactory.parseResourcesAnySyntax("FileConnectorTest/state.conf")
+        .withValue("filterOptions.sendTombstones", ConfigValueFactory.fromAnyRef("true"));
+    new FileConnector(config);
+  }
+
+  @Test
+  public void testSendTombstonesFalseDoesNotRequireIncrementalMode() throws Exception {
+    Config config = ConfigFactory.parseResourcesAnySyntax("FileConnectorTest/state.conf")
+        .withValue("filterOptions.publishMode", ConfigValueFactory.fromAnyRef("full"))
+        .withValue("filterOptions.sendTombstones", ConfigValueFactory.fromAnyRef("false"));
+    new FileConnector(config);
+  }
+
+  // Verifies that a second incremental run with no file changes produces zero docs and zero tombstones.
+  @Test
+  public void testTraversalWithStateAndMultiplePathsIncremental() throws Exception {
+    // For each archive that passes the encountered step but fails includeFile, markAllEntriesEncountered is called so their DB
+    // entries are flagged as seen. Without this, every archive entry would appear expired and generate a tombstone despite the
+    // files still existing.
+    Config config = ConfigFactory.parseResourcesAnySyntax("FileConnectorTest/stateMultiplePaths.conf")
+        .withValue("filterOptions.publishMode", ConfigValueFactory.fromAnyRef("incremental"))
+        .withValue("filterOptions.sendTombstones", ConfigValueFactory.fromAnyRef("true"));
+
+    TestMessenger messenger = new TestMessenger();
+    Publisher publisher = new PublisherImpl(config, messenger, "run", "pipeline1");
+
+    Connector connector = new FileConnector(config);
+    connector.execute(publisher);
+    assertEquals(21, messenger.getDocsSentForProcessing().size());
+
+    messenger = new TestMessenger();
+    publisher = new PublisherImpl(config, messenger, "run", "pipeline1");
+
+    // Second run: nothing changed, so nothing is republished and no tombstones are generated.
+    connector.execute(publisher);
+    List<Document> secondRunDocs = messenger.getDocsSentForProcessing();
+    long tombstoneCount = secondRunDocs.stream()
+        .filter(doc -> Boolean.TRUE.equals(doc.getBoolean(FileConnector.EXPIRED)))
+        .count();
+    assertEquals(0, secondRunDocs.size());
+    assertEquals(0, tombstoneCount);
+  }
+
+  // Verifies that a traversal with a state db and using multiple paths behaves as expected with full mode.
+  @Test
+  public void testTraversalWithStateAndMultiplePaths() throws Exception {
+    Config config = ConfigFactory.parseResourcesAnySyntax("FileConnectorTest/stateMultiplePaths.conf");
+
+    TestMessenger messenger = new TestMessenger();
+    Publisher publisher = new PublisherImpl(config, messenger, "run", "pipeline1");
+
+    Connector connector = new FileConnector(config);
+    connector.execute(publisher);
+    assertEquals(21, messenger.getDocsSentForProcessing().size());
+
+    // second run in default full mode republishes everything.
+    messenger = new TestMessenger();
+    publisher = new PublisherImpl(config, messenger, "run", "pipeline1");
+
+    connector.execute(publisher);
+    assertEquals(21, messenger.getDocsSentForProcessing().size());
+  }
+
+  // Verifies that full mode does not emit tombstones even when a previously tracked path disappears between runs.
+  @Test
+  public void testTraversalWithPathRemovedFullModeDoesNotPublishTombstones() throws Exception {
+    // Run 1 uses stateMultiplePaths.conf (example/ + defaults.csv = 21 docs). Run 2 uses state.conf
+    // (example/ only). defaults.csv is now encountered=false in the DB, but since neither config
+    // sets sendTombstones, no tombstone is generated. Run 2 publishes 18 docs (example/ in full mode).
+    Config configWithTwoPaths = ConfigFactory.parseResourcesAnySyntax("FileConnectorTest/stateMultiplePaths.conf");
+    Config configSinglePath = ConfigFactory.parseResourcesAnySyntax("FileConnectorTest/state.conf");
+
+    TestMessenger messenger1 = new TestMessenger();
+    Publisher publisher1 = new PublisherImpl(configWithTwoPaths, messenger1, "run", "pipeline1");
+    Connector connector1 = new FileConnector(configWithTwoPaths);
+    connector1.execute(publisher1);
+    assertEquals(21, messenger1.getDocsSentForProcessing().size());
+
+    TestMessenger messenger2 = new TestMessenger();
+    Publisher publisher2 = new PublisherImpl(configSinglePath, messenger2, "run", "pipeline1");
+    Connector connector2 = new FileConnector(configSinglePath);
+    connector2.execute(publisher2);
+
+    List<Document> secondRunDocs = messenger2.getDocsSentForProcessing();
+    long tombstoneCount = secondRunDocs.stream()
+        .filter(doc -> Boolean.TRUE.equals(doc.getBoolean(FileConnector.EXPIRED)))
+        .count();
+    assertEquals(18, secondRunDocs.size());
+    assertEquals(0, tombstoneCount);
+  }
+
+  // Verifies that when sendTombstones is false, deleting a file between runs does not produce tombstones.
+  @Test
+  public void testTraversalWithIncrementalTombstonesOff() throws Exception {
+    // Creates a temp directory with three files, deletes one, and confirms
+    // zero docs on the second run (no tombstone for the deleted file, no republish for unchanged files).
+    File tempDir = new File("temp-tombstones-off");
+    assertFalse(tempDir.exists());
+    tempDir.mkdir();
+
+    File file1 = new File(tempDir, "file1.txt");
+    File file2 = new File(tempDir, "file2.txt");
+    File file3 = new File(tempDir, "file3.txt");
+
+    Files.writeString(file1.toPath(), "Content 1");
+    Files.writeString(file2.toPath(), "Content 2");
+    Files.writeString(file3.toPath(), "Content 3");
+
+    try {
+      Config config = ConfigFactory.parseResourcesAnySyntax("FileConnectorTest/stateMultiplePaths.conf")
+          .withValue("paths", ConfigValueFactory.fromIterable(List.of(tempDir.toURI().toString())))
+          .withValue("filterOptions.publishMode", ConfigValueFactory.fromAnyRef("incremental"))
+          .withValue("filterOptions.sendTombstones", ConfigValueFactory.fromAnyRef("false"));
+
+      TestMessenger messenger1 = new TestMessenger();
+      Publisher publisher1 = new PublisherImpl(config, messenger1, "run1", "pipeline1");
+      Connector connector = new FileConnector(config);
+
+      // First run: all 3 files published
+      connector.execute(publisher1);
+      assertEquals(3, messenger1.getDocsSentForProcessing().size());
+
+      // Delete file2 between runs
+      Files.delete(file2.toPath());
+
+      // Second run: file2 was deleted, but sendTombstones is false so no tombstone should be published
+      // file1 and file3 are unchanged, so they are not republished either
+      TestMessenger messenger2 = new TestMessenger();
+      Publisher publisher2 = new PublisherImpl(config, messenger2, "run2", "pipeline1");
+      Connector connector2 = new FileConnector(config);
+      connector2.execute(publisher2);
+
+      List<Document> secondRunDocs = messenger2.getDocsSentForProcessing();
+      assertEquals(0, secondRunDocs.size());
+    } finally {
+      FileUtils.deleteDirectory(tempDir);
+    }
+  }
+
+  // Verifies that files generate tombstones and modified files are re-published in the same incremental run (when configured). Uses
+  // an empty state to trigger a database being created on disk to replicate a production environment
+  @Test
+  public void testTombstonesWithFileSystemChanges() throws Exception {
+
+    File tempDir = new File("temp-tombstone");
+    assertFalse(tempDir.exists());
+    tempDir.mkdir();
+
+    File file1 = new File(tempDir, "file1.txt");
+    File file2 = new File(tempDir, "file2.txt");
+    File file3 = new File(tempDir, "file3.txt");
+
+    Files.writeString(file1.toPath(), "Content 1");
+    Files.writeString(file2.toPath(), "Content 2");
+    Files.writeString(file3.toPath(), "Content 3");
+
+    try {
+      // Config with state and incremental mode with tombstones enabled
+      Config config = ConfigFactory.parseResourcesAnySyntax("FileConnectorTest/emptyState.conf")
+          .withValue("paths", ConfigValueFactory.fromIterable(List.of(tempDir.toURI().toString())))
+          .withValue("filterOptions.publishMode", ConfigValueFactory.fromAnyRef("incremental"))
+          .withValue("filterOptions.sendTombstones", ConfigValueFactory.fromAnyRef("true"));
+
+      TestMessenger messenger1 = new TestMessenger();
+      Publisher publisher1 = new PublisherImpl(config, messenger1, "run1", "pipeline1");
+      Connector connector = new FileConnector(config);
+
+      // First run, publish all 3 files
+      connector.execute(publisher1);
+      assertEquals(3, messenger1.getDocsSentForProcessing().size());
+
+      // Verify no tombstones on first run
+      long firstRunTombstones = messenger1.getDocsSentForProcessing().stream()
+          .filter(doc -> Boolean.TRUE.equals(doc.getBoolean(FileConnector.EXPIRED)))
+          .count();
+      assertEquals(0, firstRunTombstones);
+
+      // Delete file2, modify file1, leave file3 alone.
+      // Set file1's modification time explicitly to a point after the first run's traversalInstant just in case
+      Files.delete(file2.toPath());
+      Files.writeString(file1.toPath(), "Modified Content 1");
+      Files.setLastModifiedTime(file1.toPath(), java.nio.file.attribute.FileTime.from(Instant.now().plusSeconds(5)));
+
+      // Second run, should publish modified file1 + tombstone for deleted file2
+      Connector connector2 = new FileConnector(config);
+      TestMessenger messenger2 = new TestMessenger();
+      Publisher publisher2 = new PublisherImpl(config, messenger2, "run2", "pipeline1");
+      connector2.execute(publisher2);
+
+      List<Document> secondRunDocs = messenger2.getDocsSentForProcessing();
+      assertEquals(2, secondRunDocs.size());  // 1 modified file + 1 tombstone
+
+      // Verify 1 tombstone
+      long tombstoneCount = secondRunDocs.stream()
+          .filter(doc -> Boolean.TRUE.equals(doc.getBoolean(FileConnector.EXPIRED)))
+          .count();
+      assertEquals(1, tombstoneCount);
+
+      // Verify tombstone is for file2
+      Document tombstone = secondRunDocs.stream()
+          .filter(doc -> Boolean.TRUE.equals(doc.getBoolean(FileConnector.EXPIRED)))
+          .findFirst()
+          .orElseThrow();
+      assertTrue(tombstone.getString(FileConnector.FILE_PATH).contains("file2.txt"));
+      assertTrue(tombstone.isSkipped());
+
+      // Verify non-tombstone is for modified file1
+      Document modifiedDoc = secondRunDocs.stream()
+          .filter(doc -> !Boolean.TRUE.equals(doc.getBoolean(FileConnector.EXPIRED)))
+          .findFirst()
+          .orElseThrow();
+      assertTrue(modifiedDoc.getString(FileConnector.FILE_PATH).contains("file1.txt"));
+
+    } finally {
+      FileUtils.deleteDirectory(tempDir);
+
+      // Needed since our db is on disk
+      FileUtils.deleteDirectory(new File("state"));
+    }
+  }
+
 }
